Moved test for RLScheduler
